requirements.txt
langchain
langchain-community
openai
pandas
tavily-python
python-dotenv